Refactor 311 tests to share functionality
